[ADD] l10n_es_account_bank_statement_import_n43: Sustituye a l10n_es_bank_statement para importar extractos bancarios N43
