+++ .travis.yml
@@ -17,6 +17,7 @@ install:
   - git clone https://github.com/OCA/partner-contact ${HOME}/partner-contact -b ${VERSION}
   - git clone https://github.com/OCA/account-financial-tools ${HOME}/account-financial-tools -b ${VERSION}
   - git clone https://github.com/OCA/reporting-engines ${HOME}/reporting-engines -b ${VERSION}
+  - git clone https://github.com/OCA/bank-statement-import ${HOME}/bank-statement-import -b ${VERSION}
   - git clone https://github.com/OCA/maintainer-quality-tools.git ${HOME}/maintainer-quality-tools
   - export PATH=${HOME}/maintainer-quality-tools/travis:${PATH}
   - travis_install_nightly
