--- __unported__/l10n_es_bank_statement/data/account_statement_completion_rule_data.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data noupdate="1">
-
-    <record id="bank_statement_completion_rule_caixabank" model="account.statement.completion.rule">
-         <field name="name">Casar acorde a las reglas C43 de CaixaBank</field>
-         <field name="sequence">10</field>
-         <field name="function_to_call">get_from_caixabank_rules</field>
-     </record>
-
-    <record id="bank_statement_completion_rule_generic" model="account.statement.completion.rule">
-         <field name="name">Casar acorde a las reglas genéricas C43</field>
-         <field name="sequence">99</field>
-         <field name="function_to_call">get_from_generic_c43_rules</field>
-     </record>
-
-</data>
-</openerp>
