--- __unported__/l10n_es_bank_statement/i18n/l10n_es_bank_statement.pot
@@ -1,222 +1,0 @@
-# Translation of OpenERP Server.
-# This file contains the translation of the following modules:
-#	* l10n_es_bank_statement
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: OpenERP Server 7.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2014-06-10 14:31+0000\n"
-"PO-Revision-Date: 2014-06-10 14:31+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/c43_account_statement_profile.py:33
-#, python-format
-msgid "AEB C43 standard"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Amortización préstamo"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Anulación/Corrección"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:_description:0
-#: model:ir.model,name:l10n_es_bank_statement.model_account_bank_statement_line
-#, python-format
-msgid "Bank Statement Line"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: field:account.bank.statement.line,c43_concept:0
-msgid "C43 concept"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Cajero automático"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Cheque gasolina"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Compra/Venta valores"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:128
-#, python-format
-msgid "Credit amount doesn't match with the defined in the last record of account."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:124
-#, python-format
-msgid "Debit amount doesn't match with the defined in the last record of account."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Devolución/Impagado"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Div. cupones/Prima junta/Amortización"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Docimiliado/Recibo/Letra/Pago por su cta."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:115
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:119
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:123
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:127
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:136
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:147
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:194
-#, python-format
-msgid "Error in C43 file"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:137
-#, python-format
-msgid "Final balance amount = (initial balance + credit - debit) doesn't match with the defined in the last record of account."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/account_statement_completion_rule.py:33
-#, python-format
-msgid "From CaixaBank C43 rules"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/account_statement_completion_rule.py:34
-#, python-format
-msgid "From generic C43 rules"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Giro/Transferencia/Traspaso/Cheque"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Ingreso/Entrega/Abonaré"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Intereses/Comisión/Custodia/Gastos/Impuestos"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/account_statement_completion_rule.py:68
-#, python-format
-msgid "Line named \"%s\" (Ref: %s) was matched by more than one partner for VAT number \"%s\"."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/account_statement_completion_rule.py:123
-#, python-format
-msgid "Line named \"%s\" (Ref: %s) was matched by more than one partner for invoice amount \"%s\"."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:120
-#, python-format
-msgid "Number of credit records doesn't match with the defined in the last record of account."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:116
-#, python-format
-msgid "Number of debit records doesn't match with the defined in the last record of account."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:148
-#, python-format
-msgid "Number of records doesn't match with the defined in the last record."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Nómina/Seguro social"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Operaciones extranjero"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:addons/l10n_es_bank_statement/models/c43_parser.py:195
-#, python-format
-msgid "Record type %s is not valid."
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Reintegro/Talón"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Remesa efectos"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:_description:0
-#: model:ir.model,name:l10n_es_bank_statement.model_account_statement_profile
-#, python-format
-msgid "Statement Profile"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Subscripción/Div. pasivos/Canje"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Tarjeta crédito/débito"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Timbre/Corretaje/Póliza"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: selection:account.bank.statement.line,c43_concept:0
-msgid "Varios"
-msgstr ""
-
-#. module: l10n_es_bank_statement
-#: code:_description:0
-#: model:ir.model,name:l10n_es_bank_statement.model_account_statement_completion_rule
-#, python-format
-msgid "account.statement.completion.rule"
-msgstr ""
-
