--- __unported__/l10n_es_bank_statement/models/__init__.py
@@ -1,26 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright (c) All rights reserved:
-#        2013-2014 Servicios Tecnológicos Avanzados (http://serviciosbaeza.com)
-#                  Pedro Manuel Baeza <pedro.baeza@serviciosbaeza.com>
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as published
-#    by the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-from . import account_bank_statement
-from . import account_statement_profile
-from . import account_statement_completion_rule
-from . import c43_parser
-from . import c43_account_statement_profile
