--- __unported__/l10n_es_bank_statement/static/description
@@ -1,1 +1,0 @@
-src/img
