--- __unported__/l10n_es_bank_statement/views/account_bank_statement_view.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data>
-
-    <record id="view_bank_statement_form_c43" model="ir.ui.view">
-        <field name="name">account.bank.statement.c43.form</field>
-        <field name="model">account.bank.statement</field>
-        <field name="inherit_id" ref="account.view_bank_statement_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='line_ids']/tree//field[@name='date']" position="after">
-                <field name="c43_concept"/>
-            </xpath>
-            <xpath expr="//field[@name='line_ids']/form//field[@name='date']" position="after">
-                <field name="c43_concept"/>
-            </xpath>
-        </field>
-    </record>
-
-</data>
-</openerp>
