+++ l10n_es_account_bank_statement_import_n43/__openerp__.py
@@ -1,9 +1,8 @@
-# -*- coding: utf-8 -*-
+# -*- encoding: utf-8 -*-
 ##############################################################################
 #
-#    Copyright (c) All rights reserved:
-#        2013-2014 Servicios Tecnológicos Avanzados (http://serviciosbaeza.com)
-#                  Pedro Manuel Baeza <pedro.baeza@serviciosbaeza.com>
+#    Copyright(c) 2013-2015 Serv. Tecnol. Avanzados (http://serviciosbaeza.com)
+#                           Pedro M. Baeza <pedro.baeza@serviciosbaeza.com>
 #
 #    This program is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU Affero General Public License as published
@@ -19,4 +18,16 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
-from . import models
+{
+    'name': 'Importación de extractos bancarios españoles (Norma 43)',
+    'category': 'Accounting & Finance',
+    'version': '4.0',
+    'license': 'AGPL-3',
+    'author': 'Serv. Tecnol. Avanzados - Pedro M. Baeza',
+    "website": "https://github.com/OCA/l10n-spain",
+    'depends': [
+        'account_bank_statement_import',
+    ],
+    'data': [],
+    'installable': True,
+}
