+++ l10n_es_account_bank_statement_import_n43/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- encoding: utf-8 -*-
+##############################################################################
+# For copyright and license notices, see __openerp__.py file in root directory
+##############################################################################
+from . import account_bank_statement_import_n43
